Extract venue actions into useAdminVenueActions
Move all venue/layout/room/table/area mutation handlers out of AdminDashboard into a new hook (frontend/src/hooks/useAdminVenueActions.ts). AdminDashboard now imports useAdminVenueActions and removes numerous inline handlers and unused imports (adminApi mappers, layout utils). The new hook uses useVenueMutations and queryClient to update caches and preserves existing behavior (including cascade delete and area resize clamping). This centralizes venue-related logic and simplifies the AdminDashboard component.
